Standardize on uint32 for Trimesh
This includes updating the tests/the requires in Trimesh
so that the normalize.pyx file will only ever expect uint32

+++ menpo/shape/mesh/base.py
@@ -53,9 +53,10 @@ class TriMesh(PointCloud):
             if not trilist.flags.c_contiguous:
                 warn('The copy flag was NOT honoured. A copy HAS been made. '
                      'Please ensure the data you pass is C-contiguous.')
-                trilist = np.array(trilist, copy=True, order='C')
+                trilist = np.array(trilist, copy=True, order='C',
+                                   dtype=np.uint32)
         else:
-            trilist = np.array(trilist, copy=True, order='C')
+            trilist = np.array(trilist, copy=True, order='C', dtype=np.uint32)
         self.trilist = trilist
 
     def __str__(self):
